backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
pydantic[email]
flask
flask-cors
openai